@@ -1,27 +1,27 @@
 class TbDistribuicaoQuantitativoAcessosFisicosServico {
-  int idApp;
-  int index;
-  String idFichaSiciApp;
-  String id;
-  String cod_ibge;
-  String id_uf;
-  String id_municipio;
-  String id_tecnologia;
-  String pf_0;
-  String pf_512;
-  String pf_2;
-  String pf_12;
-  String pf_34;
-  String pj_0;
-  String pj_512;
-  String pj_2;
-  String pj_12;
-  String pj_34;
-  String id_lancamento;
-  String id_usuario_ultima_alteracao;
-  String municipio;
-  String uf;
-  String tecnologia;
+  int? idApp;
+  int? index;
+  String? idFichaSiciApp;
+  String? id;
+  String? cod_ibge;
+  String? id_uf;
+  String? id_municipio;
+  String? id_tecnologia;
+  String? pf_0;
+  String? pf_512;
+  String? pf_2;
+  String? pf_12;
+  String? pf_34;
+  String? pj_0;
+  String? pj_512;
+  String? pj_2;
+  String? pj_12;
+  String? pj_34;
+  String? id_lancamento;
+  String? id_usuario_ultima_alteracao;
+  String? municipio;
+  String? uf;
+  String? tecnologia;
 
   TbDistribuicaoQuantitativoAcessosFisicosServico(
       {this.idApp =0,
