@@ -17,6 +17,8 @@ import 'package:scm_engenharia_app/models/operacao.dart';
 import 'package:scm_engenharia_app/models/variaveis_de_ambiente.dart';
 import 'package:scm_engenharia_app/pages/login_page.dart';
 
+import 'help_pages/global_scaffold.dart';
+
 class PerfilPage extends StatefulWidget {
   @override
   _PerfilPageState createState() => _PerfilPageState();
@@ -25,8 +27,8 @@ class PerfilPage extends StatefulWidget {
 class _PerfilPageState extends State<PerfilPage> {
   TbUsuario _Usuariodb = new TbUsuario();
   ServicoMobileService _RestWebService = new ServicoMobileService();
-  BuildContext dialogContext;
-  DBHelper dbHelper;
+  late BuildContext dialogContext;
+  late DBHelper dbHelper;
 
   TextEditingController _TxtControllerNome = TextEditingController();
   TextEditingController _TxtControllerCpf =
@@ -39,42 +41,32 @@ class _PerfilPageState extends State<PerfilPage> {
   TextEditingController _TxtControllerEmpresa = TextEditingController();
   TextEditingController _TxtControllerUf = TextEditingController();
 
-  UF UfSelecionada;
-  List<UF> ListaUf = new List<UF>();
+  late UF UfSelecionada;
+  List<UF> ListaUf = [];
 
   Future<Null> OnGetUfs() async {
     try {
       var connectivityResult = await (Connectivity().checkConnectivity());
       if (connectivityResult == ConnectivityResult.none) {
       } else {
-        OnRealizandoOperacao("Realizando operação");
+        onRealizandoOperacao('Realizando operação ... ', true, context);
         Operacao _RestWeb = await _RestWebService.OnVariaveisDeAmbiente();
         if (_RestWeb.erro)
-          throw (_RestWeb.mensagem);
+          throw (_RestWeb.mensagem!);
         else if (_RestWeb.resultado == null)
-          throw (_RestWeb.mensagem);
+          throw (_RestWeb.mensagem!);
         else {
           VariaveisDeAmbienteResultado _Resultado =
               _RestWeb.resultado as VariaveisDeAmbienteResultado;
           setState(() {
-            ListaUf = _Resultado.uF;
+            ListaUf = _Resultado.uF!;
           });
-          if (dialogContext != null) {
-            Navigator.pop(dialogContext);
-            setState(() {
-              dialogContext = null;
-            });
-          }
+          onRealizandoOperacao('', false, context);
         }
       }
     } catch (error) {
-      if (dialogContext != null) {
-        Navigator.pop(dialogContext);
-        setState(() {
-          dialogContext = null;
-        });
-      }
-      OnToastInformacao(error);
+      onRealizandoOperacao('', false, context);
+      onAlertaInformacaoErro(error.toString(), context);
     }
   }
 
@@ -106,31 +98,22 @@ class _PerfilPageState extends State<PerfilPage> {
             _TxtControllerTelefoneWhatsapp.text;
         _ModelDadosUsuario.empresa = _TxtControllerEmpresa.text;
         _ModelDadosUsuario.uf = UfSelecionada.id;
-        OnRealizandoOperacao("Realizando cadastro");
+        onRealizandoOperacao('Realizando operação ... ', true, context);
         Operacao _RestWeb =
             await _RestWebService.OnCadastraUsuario(_ModelDadosUsuario);
         if (_RestWeb.erro)
-          throw (_RestWeb.mensagem);
+          throw (_RestWeb.mensagem!);
         else if (_RestWeb.resultado == null)
-          throw (_RestWeb.mensagem);
+          throw (_RestWeb.mensagem!);
         else {
-          if (dialogContext != null) {
-            Navigator.pop(dialogContext);
-            setState(() {
-              dialogContext = null;
-            });
-          }
-          OnToastInformacao(_RestWeb.mensagem);
+
+          onRealizandoOperacao('', false, context);
+
         }
       }
     } catch (error) {
-      if (dialogContext != null) {
-        Navigator.pop(dialogContext);
-        setState(() {
-          dialogContext = null;
-        });
-      }
-      OnToastInformacao(error);
+      onRealizandoOperacao('', false, context);
+      onAlertaInformacaoErro(error.toString(), context);
     }
   }
 
@@ -138,7 +121,7 @@ class _PerfilPageState extends State<PerfilPage> {
     try {
       Operacao _UsuarioLogado = await dbHelper.onSelecionarUsuario();
       if (_UsuarioLogado.erro)
-        throw (_UsuarioLogado.mensagem);
+        throw (_UsuarioLogado.mensagem!);
       else if (_UsuarioLogado.resultado == null) {
         Navigator.of(context).pushAndRemoveUntil(
             new MaterialPageRoute(
@@ -146,12 +129,12 @@ class _PerfilPageState extends State<PerfilPage> {
             (Route<dynamic> route) => false);
       } else {
         _Usuariodb = _UsuarioLogado.resultado as TbUsuario;
-        _TxtControllerNome.text = _Usuariodb.nome;
-        _TxtControllerCpf.text = _Usuariodb.cpf;
-        _TxtControllerEmail.text = _Usuariodb.email;
-        _TxtControllerTelefone.text = _Usuariodb.telefone;
-        _TxtControllerTelefoneWhatsapp.text = _Usuariodb.telefone;
-        _TxtControllerEmpresa.text = _Usuariodb.empresa;
+        _TxtControllerNome.text = _Usuariodb.nome!;
+        _TxtControllerCpf.text = _Usuariodb.cpf!;
+        _TxtControllerEmail.text = _Usuariodb.email!;
+        _TxtControllerTelefone.text = _Usuariodb.telefone!;
+        _TxtControllerTelefoneWhatsapp.text = _Usuariodb.telefone!;
+        _TxtControllerEmpresa.text = _Usuariodb.empresa!;
       }
 
       // OnGetUfs();
@@ -164,136 +147,7 @@ class _PerfilPageState extends State<PerfilPage> {
     }
   }
 
-  OnRealizandoOperacao(String txtInformacao) {
-    showDialog(
-      context: context,
-      barrierDismissible: false,
-      builder: (BuildContext context) {
-        dialogContext = context;
-        return Dialog(
-          child: new Row(
-            mainAxisAlignment: MainAxisAlignment.center,
-            crossAxisAlignment: CrossAxisAlignment.center,
-            mainAxisSize: MainAxisSize.max,
-            children: [
-              Container(
-                margin: const EdgeInsets.only(
-                    left: 10.0, top: 20.0, bottom: 20.0, right: 10.0),
-                child: Theme(
-                  data: Theme.of(context).copyWith(
-                    accentColor: Color(0xff018a8a),
-                  ),
-                  child: new CircularProgressIndicator(),
-                ),
-              ),
-              Expanded(
-                child: Container(
-                  margin: EdgeInsets.only(
-                      left: 10.0, top: 20.0, bottom: 20.0, right: 5.0),
-                  child: Text(
-                    txtInformacao,
-                    softWrap: true,
-                    maxLines: 1,
-                    overflow: TextOverflow.ellipsis,
-                    style: TextStyle(
-                        fontSize: 17.0,
-                        color: Color(0xff212529),
-                        fontFamily: "open-sans-regular"),
-                  ),
-                ),
-              ),
-            ],
-          ),
-        );
-      },
-    );
-  }
 
-  void OnToastInformacao(String Mensagem) {
-    showDialog(
-      context: context,
-      barrierDismissible: false,
-      builder: (BuildContext context) {
-        return Dialog(
-          shape: RoundedRectangleBorder(
-              borderRadius: BorderRadius.all(Radius.circular(8.0))),
-          child: Column(
-            mainAxisAlignment: MainAxisAlignment.center,
-            crossAxisAlignment: CrossAxisAlignment.center,
-            mainAxisSize: MainAxisSize.min,
-            children: [
-              SizedBox(height: 15.0),
-              Column(
-                mainAxisAlignment: MainAxisAlignment.center,
-                crossAxisAlignment: CrossAxisAlignment.center,
-                mainAxisSize: MainAxisSize.min,
-                children: [
-                  Text(
-                    "Informação",
-                    style: TextStyle(
-                        fontSize: 20.0,
-                        color: Color(0xff212529),
-                        fontFamily: "avenir-lt-std-roman"),
-                  ),
-                  SizedBox(
-                    height: 10,
-                  ),
-                  Divider(
-                    color: Colors.black12,
-                  ),
-                  Padding(
-                    padding: EdgeInsets.fromLTRB(15.0, 10.0, 15.0, 10.0),
-                    child: Text(
-                      Mensagem,
-                      overflow: TextOverflow.ellipsis,
-                      maxLines: 4,
-                      softWrap: false,
-                      style: TextStyle(
-                          fontSize: 17.0,
-                          color: Color(0xff212529),
-                          fontFamily: "avenir-lt-std-roman"),
-                    ),
-                  ),
-                ],
-              ),
-              Divider(
-                color: Colors.black12,
-              ),
-              Container(
-                margin: EdgeInsets.fromLTRB(0.0, 10.0, 0.0, 15.0),
-                child: new Row(
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  crossAxisAlignment: CrossAxisAlignment.center,
-                  mainAxisSize: MainAxisSize.max,
-                  children: <Widget>[
-                    FlatButton(
-                      color: Color(0xff018a8a),
-                      //`Icon` to display
-                      child: Text(
-                        '           OK           ',
-                        style: TextStyle(
-                            fontSize: 17.0,
-                            color: Color(0xffFFFFFF),
-                            fontFamily: "avenir-lt-std-roman"),
-                      ),
-                      //`Text` to display
-                      onPressed: () {
-                        Navigator.pop(context);
-                        FocusManager.instance.primaryFocus.unfocus();
-                      },
-                      shape: new RoundedRectangleBorder(
-                        borderRadius: new BorderRadius.circular(5.0),
-                      ),
-                    ),
-                  ],
-                ),
-              ),
-            ],
-          ),
-        );
-      },
-    );
-  }
 
   @override
   void initState() {
@@ -448,7 +302,7 @@ class _PerfilPageState extends State<PerfilPage> {
                                   return new DropdownMenuItem<UF>(
                                     value: value,
                                     child: Text(
-                                      value.uf,
+                                      value.uf!,
                                       textAlign: TextAlign.center,
                                       style: TextStyle(
                                           fontSize: 19.0,
@@ -458,10 +312,10 @@ class _PerfilPageState extends State<PerfilPage> {
                                     ),
                                   );
                                 }).toList(),
-                                onChanged: (UF newValue) {
+                                onChanged: (UF? newValue) {
                                   setState(() {
-                                    UfSelecionada = newValue;
-                                    _TxtControllerUf.text = newValue.uf;
+                                    UfSelecionada = newValue!;
+                                    _TxtControllerUf.text = newValue.uf!;
                                   });
                                 },
                               ),
