@@ -8,12 +8,12 @@ import 'package:scm_engenharia_app/pages/erro_informacao_page.dart';
 
 class ComponentsJWTToken {
 
-  static Future<String> JWTTokenPadrao() async {
+  static Future<String?> JWTTokenPadrao() async {
     try {
       DBHelper  dbHelper = DBHelper();
       Operacao _UsuarioLogado = await dbHelper.onSelecionarUsuario();
       if (_UsuarioLogado.erro)
-        throw (_UsuarioLogado.mensagem);
+        throw (_UsuarioLogado.mensagem!);
       else if (_UsuarioLogado.resultado == null) {
         return Components.JWTTokenPadrao();
       }
@@ -22,17 +22,18 @@ class ComponentsJWTToken {
         String key =
             "bc47f175a831996b652146d47e159349f75e6c4665570ef35606678a18054d13";
         final claimSet = new JwtClaim(otherClaims: <String, Object>{
-          "user": "" + _Usuariodb.email + "",
-          "pass": "" + _Usuariodb.senha + "",
+          "user": "" + _Usuariodb.email! + "",
+          "pass": "" + _Usuariodb.senha! + "",
         });
         // Generate a JWT from the claim set
         final token = issueJwtHS256(claimSet, key);
         return token;
       }
-      return null;
+
     } catch (error) {
 
     }
+    return '';
   }
 
 }
