@@ -1,4 +1,4 @@
 library UsuarioLogado;
 import 'package:scm_engenharia_app/data/tb_usuario.dart';
 
-TbUsuario DadosUsuarioLogado = null;
+TbUsuario? DadosUsuarioLogado = null;
